Version 2.1.0 Including: - Replace currency from cents to full using floats - Added custom types for retrieving StripesPerMonth - Added cost per stripe for day

+++ src/main/java/nl/brandfonds/Brandfonds/model/Receipt.java
@@ -23,9 +23,9 @@ public class Receipt {
     private String fileName;
     private String description;
     private LocalDate paidDate;
-    private long paidAmount;
+    private Float paidAmount;
 
-    public Receipt(String filename, String description, LocalDate paidDate, long paidAmount) {
+    public Receipt(String filename, String description, LocalDate paidDate, Float paidAmount) {
         this.fileName = filename;
         this.description = description;
         this.paidDate = paidDate;
